requirements.txt
fastapi==0.103.0
uvicorn==0.23.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==0.26.0
python-dotenv==0.21.0